update: split footer component into footersections, add footer sections to app layout, style app layout with refinements
